SupportingDocs/app_hybrid_search/requirements.txt
flask
gunicorn
openai
azure-identity
azure-storage-blob
azure-search-documents
pypdf
azure-search-documents